+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:41521/be055c8e-0c29-4cc9-a18c-09a705185686?"}
+{"mongoUri":"mongodb://127.0.0.1:44979/7174e8e8-17c2-4e9e-bca0-7249e88ec70a?"}
