+++ src/main/factories/login/login-validation.ts
@@ -1,8 +1,6 @@
-import { RequiredFielValidation } from '../../../presentation/helpers/validators/required-field-validation'
+import { RequiredFielValidation, EmailValidation, ValidationComposite } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
 import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
-import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
 
 export const makeLoginValidation = (): Validation => {
   const validations: Validation[] = []
