+++ src/main/factories/signup/signup-validation.ts
@@ -1,9 +1,11 @@
-import { RequiredFielValidation } from '../../../presentation/helpers/validators/required-field-validation'
+import {
+  RequiredFielValidation,
+  CompareFieldsValidation,
+  ValidationComposite,
+  EmailValidation
+} from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
-import { CompareFieldsValidation } from '../../../presentation/helpers/validators/compare-field-validation'
 import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
-import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
 
 export const makeSignUpValidation = (): Validation => {
   const validations: Validation[] = []
